Bug 1645896 - Cast state constants to uint32_t in switch block. r=timhuang
Differential Revision: https://phabricator.services.mozilla.com/D80942

UltraBlame original commit: 2586226d649aa7ea0ff1c07b7bb1746255607350

+++ toolkit/components/antitracking/ContentBlockingNotifier.cpp
@@ -140,28 +140,33 @@ void ReportBlockingToConsole(uint64_t aWindowID, nsIURI* aURI,
         nsAutoCString category;
         
         
+        
+        
+        
         switch (aRejectedReason) {
-          case nsIWebProgressListener::STATE_COOKIES_BLOCKED_BY_PERMISSION:
+          case uint32_t(
+              nsIWebProgressListener::STATE_COOKIES_BLOCKED_BY_PERMISSION):
             message = "CookieBlockedByPermission";
             category = NS_LITERAL_CSTRING("cookieBlockedPermission");
             break;
 
-          case nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER:
+          case uint32_t(nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER):
             message = "CookieBlockedTracker";
             category = NS_LITERAL_CSTRING("cookieBlockedTracker");
             break;
 
-          case nsIWebProgressListener::STATE_COOKIES_BLOCKED_ALL:
+          case uint32_t(nsIWebProgressListener::STATE_COOKIES_BLOCKED_ALL):
             message = "CookieBlockedAll";
             category = NS_LITERAL_CSTRING("cookieBlockedAll");
             break;
 
-          case nsIWebProgressListener::STATE_COOKIES_BLOCKED_FOREIGN:
+          case uint32_t(nsIWebProgressListener::STATE_COOKIES_BLOCKED_FOREIGN):
             message = "CookieBlockedForeign";
             category = NS_LITERAL_CSTRING("cookieBlockedForeign");
             break;
 
-          case nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN:
+          case uint32_t(
+              nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN):
             message = "CookiePartitionedForeign";
             category = NS_LITERAL_CSTRING("cookiePartitionedForeign");
             break;
